# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 3D Designer API - Garden Furniture & Garden Buildings
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Settings {
  /** Base URL for local Ollama instance */
  ollamaUrl: string;
  /** Ollama model to use (e.g. qwen2.5) */
  ollamaModel: string;
  /** Base URL for Open-WebUI instance */
  openWebUiUrl: string;
}

export interface FurnitureStyle {
  id: string;
  name: string;
  description: string;
}

export interface FurnitureStyleList {
  styles: FurnitureStyle[];
}

export interface FurnitureItem {
  id: string;
  name: string;
  description: string;
  category: string;
}

export interface FurnitureItemList {
  items: FurnitureItem[];
}

export interface FurnitureGenerateRequest {
  styleId?: string | null;
  itemId?: string | null;
  /** User prompt with dimensions and design details */
  prompt: string;
  projectName?: string | null;
}

export type GenerateResponseStatus =
  (typeof GenerateResponseStatus)[keyof typeof GenerateResponseStatus];

export const GenerateResponseStatus = {
  queued: "queued",
  processing: "processing",
  rendering: "rendering",
  complete: "complete",
  error: "error",
} as const;

export interface GenerateResponse {
  jobId: string;
  status: GenerateResponseStatus;
  /** Current stage description */
  stage: string;
  /** The generated script/model output */
  modelOutput?: string | null;
  estimatedSeconds?: number | null;
  error?: string | null;
}

export type ProjectType = (typeof ProjectType)[keyof typeof ProjectType];

export const ProjectType = {
  furniture: "furniture",
  building: "building",
} as const;

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  queued: "queued",
  processing: "processing",
  rendering: "rendering",
  complete: "complete",
  error: "error",
} as const;

export interface Project {
  id: string;
  name: string;
  type: ProjectType;
  styleId?: string | null;
  itemId?: string | null;
  prompt: string;
  status: ProjectStatus;
  createdAt: string;
  updatedAt: string;
}

export type CreateProjectRequestType =
  (typeof CreateProjectRequestType)[keyof typeof CreateProjectRequestType];

export const CreateProjectRequestType = {
  furniture: "furniture",
  building: "building",
} as const;

export interface CreateProjectRequest {
  name: string;
  type: CreateProjectRequestType;
  styleId?: string | null;
  itemId?: string | null;
  prompt: string;
}

export interface ProjectList {
  projects: Project[];
}

export interface ErrorResponse {
  error: string;
  message: string;
}