# lib/api-zod/src/generated/types/createProjectRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 3D Designer API - Garden Furniture & Garden Buildings
 * OpenAPI spec version: 0.1.0
 */
import type { CreateProjectRequestType } from "./createProjectRequestType";

export interface CreateProjectRequest {
  name: string;
  type: CreateProjectRequestType;
  styleId?: string | null;
  itemId?: string | null;
  prompt: string;
}