# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 3D Designer API - Garden Furniture & Garden Buildings
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectStatus } from "./projectStatus";
import type { ProjectType } from "./projectType";

export interface Project {
  id: string;
  name: string;
  type: ProjectType;
  styleId?: string | null;
  itemId?: string | null;
  prompt: string;
  status: ProjectStatus;
  createdAt: Date;
  updatedAt: Date;
}