# lib/api-zod/src/generated/types/projectList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 3D Designer API - Garden Furniture & Garden Buildings
 * OpenAPI spec version: 0.1.0
 */
import type { Project } from "./project";

export interface ProjectList {
  projects: Project[];
}